Rename `configPath` to `config` and make sure config.json is valid
Also splits index.js to make sure the bulk of it is typed

Update pinejs from 10.8.0 to 10.11.0

Change-type: minor
