+++ package.json
@@ -12,16 +12,16 @@
     "dist/"
   ],
   "scripts": {
-    "build": "npm run clean && tsc --project . && copyup \"src/**/*.sbvr\" \"src/**/*.sql\" dist/",
+    "build": "npm run clean && tsc --project ./tsconfig.build.json && copyup \"src/**/*.sbvr\" \"src/**/*.sql\" dist/",
     "clean": "rimraf dist/",
-    "lint": "resin-lint --typescript src/ && tsc --noEmit --project .",
-    "prettify": "prettier --config ./node_modules/resin-lint/config/.prettierrc --write index.js \"src/**/*.ts\" \"typings/**/*.ts\"",
+    "lint": "resin-lint --typescript src/ init.ts && tsc --noEmit --project .",
+    "prettify": "prettier --config ./node_modules/resin-lint/config/.prettierrc --write index.js init.ts \"src/**/*.ts\" \"typings/**/*.ts\"",
     "prepare": "npm run build",
     "test": "npm run build && npm run lint"
   },
   "dependencies": {
     "@resin.io/device-types": "^10.5.0",
-    "@resin/pinejs": "^10.8.0",
+    "@resin/pinejs": "^10.11.0",
     "@types/basic-auth": "^1.1.2",
     "@types/bluebird": "^3.5.25",
     "@types/cookie-session": "^2.0.36",
@@ -41,9 +41,9 @@
     "@types/request": "^2.48.1",
     "array-sort": "^1.0.0",
     "avsc": "^5.4.7",
-    "aws-sdk": "^2.429.0",
+    "aws-sdk": "^2.437.0",
     "basic-auth": "^2.0.1",
-    "bluebird": "^3.5.3",
+    "bluebird": "^3.5.4",
     "body-parser": "^1.18.3",
     "compression": "^1.7.4",
     "cookie-parser": "^1.4.4",
@@ -72,13 +72,13 @@
     "thirty-two": "^1.0.0",
     "ts-node": "^7.0.1",
     "typed-error": "^2.0.0",
-    "typescript": "^3.3.4000"
+    "typescript": "^3.4.3"
   },
   "devDependencies": {
     "copyfiles": "^2.1.0",
     "husky": "^1.3.1",
     "lint-staged": "^8.1.5",
-    "prettier": "^1.16.4",
+    "prettier": "^1.17.0",
     "resin-lint": "^2.0.0",
     "rimraf": "^2.6.3"
   },
